requirements.txt
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg==3.3.2
requests==2.32.5
